Made the move history list into its own component.
